CommonFeatures

Documentation for CommonFeatures.

CommonFeatures.art

CommonFeatures.fullinfocrit

CommonFeatures.infocrit

CommonFeatures.isstable

CommonFeatures.makecompanion

CommonFeatures.ridgerankselect

CommonFeatures.rrvar

CommonFeatures.rrvaric

CommonFeatures.simstats

CommonFeatures.simulatemardata

CommonFeatures.simulatetuckerdata

CommonFeatures.tensorols

CommonFeatures.tlag

CommonFeatures.tuckerpar

CommonFeatures.tuckerreg

CommonFeatures.art — Function

art(Y, p)

Compute the Tensor Autoregressive (ART) model using tensor ordinary least squares.

Arguments

Y::AbstractArray: Matrix or Tensor of size (n1, n2, T) or (n1, n2, ..., nk, T).

p::Int: Order of the autoregressive model.

Returns

result: Tensor of size (n1, n2, ..., nk, n1, ..., nk).

Examples

Y = rand(3, 2, 100)
p = 2

result = art(Y, p)

CommonFeatures.fullinfocrit — Function

fullinfocrit(mardata, p, r̄, maxiters, tucketa, ϵ, stdize)

Calculate information criteria (AIC and BIC) for different combinations of Tucker ranks in a matrix autoregressive (MAR) model. Iterates over different time lags to find the best Tucker ranks.

Fields

mardata::AbstractArray: The matrix-valued time series data.

pmax::Int: The maximum order of the autoregressive model to be considered.

r̄::AbstractVector: A vector specifying the maximum Tucker ranks for each mode. Default is all possible combinations.

maxiters::Int: An integer specifying the number of iterations the Tucker regression should run. Default value is 500.

tucketa::Real: A real value specifying the step size for the Tucker regression. Default value is 1e-02.

ϵ::Real: A real value specifying the convergence criterion for the Tucker regression. Default value is 1e-03.

stdize::Bool: A boolean value specifying whether to standardize the data before running the Tucker regression. Default value is false.

Output

A tuple with the following elements:

BIC: The Tucker ranks chosen based on the Bayesian Information Criterion (BIC).

AIC: The Tucker ranks chosen based on the Akaike Information Criterion (AIC).

ictable: A 6xN matrix where N is the number of valid Tucker rank combinations. Each column corresponds to a combination, and rows contain the following information:

Row 1: Log determinant of the covariance matrix plus a penalty term (BIC criterion).

Row 2: Log determinant of the covariance matrix plus a penalty term (AIC criterion).

Rows 3-6: The chosen Tucker ranks for each mode.

regiters: The number of iterations for each Tucker regression.

numconv: The number of converged Tucker regressions.

Example

mardata = randn(4,3,100)  # Example matrix time series data
p = 2
r̄ = [2, 2, 2, 2]
result = infocrit(mardata, p, r̄)
println("BIC Chosen Ranks: ", result.BIC)
println("AIC Chosen Ranks: ", result.AIC)
println("Information Criteria Table: ", result.ictable)

CommonFeatures.infocrit — Function

infocrit(mardata, p, r̄, maxiters, tucketa, ϵ, stdize)

Calculate information criteria (AIC and BIC) for different combinations of Tucker ranks in a matrix autoregressive (MAR) model.

Arguments

mardata::AbstractArray: The matrix-valued time series data.

p::Int: The order of the autoregressive model.

r̄::AbstractVector: A vector specifying the maximum Tucker ranks for each mode. Default is all possible combinations.

maxiters::Int: An integer specifying the number of iterations the Tucker regression should run. Default value is 500.

tucketa::Real: A real value specifying the step size for the Tucker regression. Default value is 1e-02.

ϵ::Real: A real value specifying the convergence criterion for the Tucker regression. Default value is 1e-03.

stdize::Bool: A boolean value specifying whether to standardize the data before running the Tucker regression. Default value is false.

Output

A tuple with the following elements:

BIC: The Tucker ranks chosen based on the Bayesian Information Criterion (BIC).

AIC: The Tucker ranks chosen based on the Akaike Information Criterion (AIC).

ictable: A 6xN matrix where N is the number of valid Tucker rank combinations. Each column corresponds to a combination, and rows contain the following information:

Row 1: Log determinant of the covariance matrix plus a penalty term (BIC criterion).

Row 2: Log determinant of the covariance matrix plus a penalty term (AIC criterion).

Rows 3-6: The chosen Tucker ranks for each mode.

Row 7: The number of iterations for each Tucker regression.

numconv: The number of converged Tucker regressions.

Example

mardata = randn(4,3,100)  # Example matrix time series data
p = 2
r̄ = [2, 2, 2, 2]
result = infocrit(mardata, p, r̄)
println("BIC Chosen Ranks: ", result.BIC)
println("AIC Chosen Ranks: ", result.AIC)
println("Information Criteria Table: ", result.ictable)

CommonFeatures.isstable — Function

isstable(var)

Check the stability of a VAR (Vector Autoregressive) model.

This function checks the stability of a VAR model by analyzing its companion matrix eigenvalues. A VAR model is considered stable if all the eigenvalues of its companion matrix are within the unit circle.

Arguments

var: Lag matrix in the form required for a VAR model. See the documentation of VAR.

Returns

Bool: Returns true if the VAR model is stable, and false otherwise.

Example

B = [1.0 2.0;
     3.0 4.0]
var_stable = isstable(B)  # Returns true or false based on the stability of the VAR model

Note

The stability of a VAR model is determined by analyzing the eigenvalues of its companion matrix. The companion matrix is constructed using the makecompanion function.

See Also

makecompanion: Function to create the VAR companion matrix.

CommonFeatures.makecompanion — Method

makecompanion(B::AbstractMatrix{T}) where {T}

Create the VAR companion matrix.

Given a VAR of the form yₜ = b₀ + B₁ yt-1 + … + Bₚyt-p + εₜ

Create the companion matrix. Thus, $B$ is a $np imes np$ matrix.

Arguments

-B::AbstractMatrix{T}: Lag matrix in the form required for a VAR model. See the documentation of VAR.

References

Kilian, L., & Lütkepohl, H. (2017). Structural Vector Autoregressive Analysis: (1st ed.). Cambridge University Press. https://doi.org/10.1017/9781108164818

CommonFeatures.ridgerankselect — Function

ridgerankselect(est::AbstractArray, c::Real, r̄=collect(size(est)))::Vector{Int}

Used to provide an initial guess of the ranks for a Tucker regression using the framework of Lam and Yao (2011) and Wang et. al (2023). Given a four-dimensional initial estimate est (typically from the art function), this function performs ridge rank selection on each mode of the tensor using a ridge parameter c. The result is a vector of integers representing the selected ranks for each mode.

Arguments

est::AbstractArray: The input four-dimensional array (tensor) for ridge rank selection.

c::Real: Ridge parameter controlling regularization strength.

r̄::Vector{Int} = collect(size(est)): Vector specifying the maximum ranks to consider for each mode. Default is the size of the corresponding mode in est.

Returns

Vector{Int}: A vector containing the selected ranks for each mode of the input tensor.

Example

est = rand(3, 4, 5, 6)
c = 0.1
selected_ranks = ridgerankselect(est, c)

References

Lam, Clifford, and Qiwei Yao. "Factor modeling for high-dimensional time series: inference for the number of factors." The Annals of Statistics (2012): 694-726.

Wang, Di, et al. "High-dimensional vector autoregressive time series modeling via tensor decomposition." Journal of the American Statistical Association 117.539 (2022): 1338-1356.

CommonFeatures.rrvar — Function

rrvar(Y::AbstractMatrix, r::Int, p::Int)

Compute reduced-rank vector autoregressive (RRVAR) model parameters.

Arguments

Y::AbstractMatrix: Input matrix with dimensions (k, T), where k is the number of variables and T is the number of observations.

r::Int: Rank of the reduced-rank model.

p::Int: Order of the autoregressive model.

Returns

C::Matrix: Reduced-rank VAR coefficient matrix.

A::Matrix: Left reduced-rank matrix.

B::Matrix: Right reduced-rank matrix.

Details

This function computes the reduced-rank VAR parameters using the method.

CommonFeatures.rrvaric — Method

rrvaric(vardata, pmax, stdize)

Compute the optimal rank of a reduced rank regression using information criteria.

Fields

vardata::AbstractMatrix: The data matrix where each column represents a variable and each row represents an observation.

pmax::Int: The maximum lag order.

stdize::Bool: Indicates whether to standardize the data or not.

Returns

A tuple with the following elements:

BIC: The optimal rank based on the Bayesian Information Criterion.

AIC: The optimal rank based on the Akaike Information Criterion.

HQ: The optimal rank based on the Hannan-Quinn Information Criterion.

ictable: A 5xN matrix where N is the number of valid rank combinations. Each column corresponds to a combination, and rows contain the following information:

Row 1: Log determinant of the covariance matrix plus a penalty term (BIC criterion).

Row 2: Log determinant of the covariance matrix plus a penalty term (AIC criterion).

Row 3: Log determinant of the covariance matrix plus a penalty term (HQ criterion).

Row 4: The chosen rank.

Row 5: The chosen lag order.

CommonFeatures.simstats — Method

simstats(selectedranks, correctrank, sims)

Calculate simulation statistics based on selected ranks, correct ranks, and the number of simulations.

Arguments

selectedranks::AbstractMatrix: Matrix of selected ranks for each simulation. Simulations should be along the rows.

correctrank::AbstractVector: Vector of correct ranks for each simulation.

sims::Int: CNumber of simulations.

Returns

A named tuple containing the following statistics:

avgrank: Average rank for each simulation.

stdrank: Standard deviation of ranks for each simulation.

freqcorrect: Frequency of correct ranks for each simulation.

freqhigh: Frequency of ranks higher than the correct rank for each simulation.

freqlow: Frequency of ranks lower than the correct rank for each simulation.

CommonFeatures.simulatemardata — Function

simulatemardata(dimvals, obs, scale)

Simulate matrix autoregressive (MAR) data with specified dimensions, observation count, and scaling factor.

Arguments

dimvals::AbstractVector: Dimensions of the MAR tensor (dimvals[1] for the first mode, dimvals[2] for the second mode).

obs::Int: Number of observations to simulate.

scale::Real: Scaling factor for the MAR data.

Returns

A named tuple containing:

mardata::Array{Float64, 3}: Simulated MAR data with dimensions (dimvals[1], dimvals[2], obs).

stabit::Int: Number of iterations required to generate a stable tensor.

A::Array{Float64, 4}: Autoregressive tensor.

Examples

result = simulatemardata([5, 4], 100, 1.0)

CommonFeatures.simulatetuckerdata — Function

simulatetuckerdata(dimvals, ranks, obs, scale)

Simulate Tucker data with specified dimensions, ranks, observation count, and scaling factor.

Arguments

dimvals::AbstractVector: Dimensions of the tensor (dimvals[1] for the first mode, dimvals[2] for the second mode).

ranks::AbstractVector: Tucker ranks for the four modes.

obs::Int: Number of observations to simulate.

A::Array{Float64, 4}: Coefficient tensor. If nothing, a random coefficient tensor will be generated.

p::Int: Number of lags to include. Default is 1 and the maximum is 5.

snr::Real: Desired signal-to-noise ratio. Default is 0.7.

Returns

A named tuple containing:

data::Array{Float64, 3}: Simulated Tucker data with dimensions (dimvals[1], dimvals[2], obs).

A::Array{Float64, 4}: Chosen coefficient tensor.

Σ::Array{Float64, 2}: Covariance matrix of the noise.

Examples

result = simulatetuckerdata([5, 4], [2, 3, 2, 3], 100, 1.0)

CommonFeatures.tensorols — Method

tensorols(Y, X)

Compute tensor ordinary least squares equation by equation.

Arguments

Y::AbstractArray: Tensor of size (n1, n2, ..., nk, T).

X::AbstractArray: Tensor of size (p1, p2, ..., pk, T).

Returns

result: Tensor of size (n1, n2, ..., nk, p1, p2, ..., pk).

CommonFeatures.tlag — Function

tlag(ten_data::AbstractArray, p=1)

Create lagged arrays from a tensor-valued time series of the form N₁ × N₂ × … × Nₖ × T. New lagged is a tensor of size N1 × N2 × … × N_k × p × T.

Arguments:

ten_data: A tensor-valued time series with dimensions N₁ × N₂ × … × Nₖ × T, where N₁, N₂, ..., Nₖ represent the spatial dimensions, and T represents the time dimension.

p: The number of lags to be created. Default is 1, indicating a single lag.

Returns:

original: The original tensor without any lag.

lagged: A tensor containing lagged versions of the original tensor, with dimensions N₁ × N₂ × … × Nₖ × p × T, where lastdim is the size of the last dimension of the input tensor.

Example:

data = randn(3, 4, 5, 10)  # Example tensor with dimensions 3 × 4 × 5 × 10
original, lagged = tlag(data, 2)  # Create lagged arrays with 2 lags

CommonFeatures.tuckerpar — Function

tuckerpar(dimvals::AbstractVector, ranks::AbstractVector, P::Integer=1)

Compute the Tucker compression parameter for a tensor with specified dimensions and Tucker ranks.

Arguments

dimvals::AbstractVector: A vector representing the dimensions of the original tensor.

ranks::AbstractVector: A vector representing the Tucker ranks for compression. It should have twice the length of dimvals.

P::Integer=1: An optional parameter representing the mode-n unfolding size (default is 1).

Output

Returns the Tucker compression parameter for the given input.

Examples

dimvals = [3, 4, 5]
ranks = [2, 3, 2, 4]
P = 2
result = tuckerpar(dimvals, ranks, P)
println(result)  # Output: 44

References

Tucker, L. R. (1966). Some mathematical notes on three-mode factor analysis. Psychometrika, 31(3), 279-311.

CommonFeatures.tuckerreg — Function

tuckerreg(mardata, ranks::AbstractVector, eta::Real, a=1, b=1, ϵ=1e-04, maxiter=1000, p=1,ϵ=1e-02)

Perform a matrix autoregression with one lag (MAR(1)) with a reduced tucker rank along the coefficients. Uses the gradient descent algorithm of Wang, Zhang, and Li 2024.

Arguments

mardata::AbstractArray: Input tensor data. Should be $N1 imes N2 imes T$

ranks::AbstractVector: Vector specifying the desired ranks of the Tucker decomposition.

eta::AbstractFloat: Learning rate for gradient descent (default: 1e-03).

a::Real: Regularization parameter (default: 1).

b::Real: Regularization parameter (default: 1).

maxiter::Int: Maximum number of iterations for gradient descent (default: 3000).

p::Int: Number of lags to include

ϵ::AbstractFloat: Convergence threshold for stopping criteria (default: 1e-03).

Returns

A tuple containing the Tucker decomposition components:

G: Core tensor.

U1, U2, U3, U4: Factor matrices.

A: Full Tucker decomposition tensor.

iters: Number of iterations performed.

fullgrads: A matrix keeping track of gradients. Can be plotted to determine whether gradients behave properly.